Fix Daily Bank Transaction

+++ Com.Danliris.Service.Finance.Accounting.Lib/BusinessLogic/Interfaces/CreditorAccount/ICreditorAccountService.cs
@@ -11,7 +11,7 @@ namespace Com.Danliris.Service.Finance.Accounting.Lib.BusinessLogic.Interfaces.C
 {
     public interface ICreditorAccountService
     {
-        (ReadResponse<CreditorAccountViewModel>, double) GetReport(int page, int size, string suplierName, int month, int year, int offSet);
+        (ReadResponse<CreditorAccountViewModel>, decimal) GetReport(int page, int size, string suplierName, int month, int year, int offSet);
         MemoryStream GenerateExcel(string suplierName, int month, int year, int offSet);
         Task<int> UpdateFromUnitPaymentOrderAsync(CreditorAccountUnitPaymentOrderPostedViewModel viewModel);
         Task<int> CreateFromUnitReceiptNoteAsync(CreditorAccountUnitReceiptNotePostedViewModel viewModel);
